Build SSL libraries statically
As posted in the forums (http://www.eternal-lands.com/forum/index.php?/topic/61508-new-development-version-of-the-android-client/&do=findComment&comment=597504)
older android versions fail to run the client because it tries to use the
system's OpenSSL library, instead of the one bundled with the client. Build
and link the OpenSSL libraries statically in an attempt to fix this.